+++ Changelog.md
@@ -6,6 +6,12 @@
    loaded from an SDF DOM object.
     * [Pull Request 302](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/302)
 
+1. Throttle sensors update rate
+    * [Pull Request 323](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/323)
+
+1. Added system for ignition::sensors::AirPressureSensor.
+    * [Pull Request 300](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/300)
+
 1. Support conversion and serialization of PBR parameters in a material component
     * [Pull Request 304](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/304)